requirements.txt
requests>=2.28
python-dotenv>=1.0.0
tqdm>=4.64
click>=8.0
numpy>=1.21.0